+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 - HTML+ERB support on README
 - BaseChart as a parent for Cartesian Chart and Polar Chart
 - OptionsBuilder tests for x and y axes
+- Option `parentHeightOffset` for `chart`
+- Option `radar` for `plot_options`
 
 ### Fixed
 - Array of hashes with key `data` as input not read correctly